change parser decl api to be nicer
using tagged template literals was mostly in use to make it distinct from the Rule.ruleName and Rule.$mod usage, but it was clunky to type them so often in a code base.
Switching to an object doesn't increase the indentation at all, but makes all of the Rule() calls look cleaner, and rules can be mixed together using objects which is kind of neat.